Implement hover range for procedure calls.
This could be made more elegant, but it would require some changes to
the Tcl parser. Currently, there is no way to distinguish between
multiple commands inside of a command word. For example, from this:

[set x world; return "hello $x"]

we would not be able to tell from the AST that there are two
procedures being invoked.