Modify SoarFile to be immutable (or as immutable as possible)
Since some of the members of SoarFile are not entirely under our
control (technically we can modify the TclAstNode class, but that
would be a larger refactor than I want to undertake right now) it
actually does expose mutable objects. I am willing to live with these
relatively weak guarantees though.

One issue I anticipate is that we have been storing the contents of
AST nodes on the nodes themselves, lazily evaluated. This would be an
issue if we are concurrently accessing the TclAstNode class.