Perform analyses in background threads.
The getAnalysis() now returns an CompletableFuture. In most cases,
this future will resolve immediately, except for when the very first
analysis is running.

The implementations of a number of queries had to be adjusted. The
diffs look large, but nothing really changed other than the fact that
the application logic has been moved into the future monad.