Move remaining diagnostics logic into the analysis class.
The main win from this is that we will compute proper diagnostics
regardless of whether a file is actually open.

The downside is that the analysis pass is a bit slower than what we
had before, so these diagnostics will be somewhat delayed.

Aside from general improvements in the performance of analysis, a
possible optimization in the future would be to publish diagnostics
before the analysis has completed.