Change internal representation of URIs from String to the URI class.
Fixes #21

The lsp4j library still uses String as its API type for URIs (although
I would argue that it should be using URI anyway), so we still convert
to/from String at the edges of the system.

This commit also intentionally calls URLEncoder inside the server test
fixture in order to simulate the percent-encoding behaviour we see in
some environments.