Use codeAction request to trigger tcl expansion refresh instead of hover.
This works much better than hover for two reasons.

The first reason is that code actions are triggered by the location of
the cursor, rather than the mouse pointer. This means that the user
has much more precise control over what gets put in the expanded
buffer.

The second reason is that the code action request contains a range
instead of just a position. This means that we can let the user select
multiple commands and show all the expansions for that selection.