Clear file contents before inserting expanded Tcl; fix for #141.
By clearing the contents first, we force clients to scroll to the top
of the buffer, so that when we later insert new text, the contents
will be in view.

The cursor itself still ends up at the end of the file; we don't
have any control over this. However, the scroll position is where we
need it to be.