Track files that are attempted to be sourced, even if unsuccessful.
This should fix #164. The issues was that creating and modifying a new
file would not trigger a new analysis, because the results of the
previous analysis did not indicate that this new file should have any
effect (since it didn't exist at the time the previous analysis was
run). This change tracks the set of URIs that are _attempted_ to be
analysed, even if the analysis in unsuccessful.